update: signInSuccess 기능 signup으로 결합
signInSuccess 기능을 signup으로 결합하고, 전자는 삭제
